+++ objects.py
@@ -2,7 +2,7 @@ from pygame.draw import circle, rect, line, polygon
 from pygame.mouse import get_pressed, get_pos
 from time import time
 import pygame
-
+from math import ceil
 
 # some mindless object templates
 
@@ -19,7 +19,6 @@ class Rect:
         return
 
     pass
-
 
 class Button:
     last_click = time()
@@ -38,8 +37,9 @@ class Button:
             self.resp()
         return
 
+    def is_clicked(self):
+        return self.check_click() and time() - self.last_click >= self.delay_time
     pass
-
 
 class Rect_Button(Button):
     def __init__(self, scr, color, x, y, w, h, resp=lambda: None, width=0, delay_time=0.3):
@@ -56,7 +56,6 @@ class Rect_Button(Button):
         return
 
     pass
-
 
 class Text:
     def __init__(self, scr, txt, x, y, font_size, color, font="freesansbold.ttf"):
@@ -92,14 +91,17 @@ class Text:
 
     pass
 
-
 class Rect_Text:
-    def __init__(self, scr, txt, t_x, t_y, size, t_color, r_color):
-        self.text = Text(scr, txt, t_x, t_y, size, t_color)
+    def __init__(self, scr, txt, x, y, font_size, t_color, r_color, scale=1.1):
+        text = Text(scr, txt, x, y, font_size, t_color)
+        x, y, w, h = text.get_rekt()
+        self.text = Text(scr, txt, x + (w * (scale - 1)) / 2, y + (h * (scale - 1)) / 2, font_size, t_color)
+        self.scaled_text = Text(scr, txt, x, y, ceil(font_size * scale), t_color)
+        x, y, w, h = self.scaled_text.get_rekt()
         self.scr = scr
         self.r_color = r_color
-        x, y, w, h = self.text.get_rekt()
         self.rect = Rect(scr, r_color, x, y, w, h)
+        self.scale = scale
         return
 
     def blit(self):
@@ -108,18 +110,23 @@ class Rect_Text:
         return
 
     def update_text(self, new):
+        self.scaled_text.update_text(new)
+        x, y, w, h = self.scaled_text.get_rekt()
         self.text.update_text(new)
-        x, y, w, h = self.text.get_rekt()
+        self.text.x = x + (w * (self.scale - 1)) / 2
+        self.text.y = y + (h * (self.scale - 0.9)) / 2
         self.rect = Rect(self.scr, self.r_color, x, y, w, h)
         return
 
     pass
 
-
 class Text_Button:
-    def __init__(self, scr, txt, x, y, font_size, t_color, r_color, font="freesansbold.ttf", resp=lambda: None):
-        self.text = Text(scr, txt, x, y, font_size, t_color, font)
-        x, y, w, h = self.text.get_rekt()
+    def __init__(self, scr, txt, x, y, font_size, t_color, r_color, font="freesansbold.ttf", resp=lambda: None, scale=1.1):
+        text = Text(scr, txt, x, y, font_size, t_color, font)
+        x, y, w, h = text.get_rekt()
+        self.text = Text(scr, txt, x + (w * (scale - 1)) / 2, y + (h * (scale - 1)) / 2, font_size, t_color, font)
+        self.scaled_text = Text(scr, txt, x, y, ceil(font_size * scale), t_color, font)
+        x, y, w, h = self.scaled_text.get_rekt()
         self.resp = resp
         self.Rect = Rect_Button(scr, r_color, x, y, w, h, resp)
         self.scr = scr
@@ -131,9 +138,71 @@ class Text_Button:
         self.text.blit()
         return
 
-    def update_text(self, new_text):
-        self.text.update_text(new_text)
-        self.Rect.rect = self.text.get_rekt()
+    def update_text(self, new):
+        self.scaled_text.update_text(new)
+        x, y, w, h = self.scaled_text.get_rekt()
+        self.text.update_text(new)
+        self.text.x = x + (w * (self.scale - 1)) / 2
+        self.text.y = y + (h * (self.scale - 0.9)) / 2
+        self.rect = Rect(self.scr, self.r_color, x, y, w, h)
         return
 
     pass
+
+class Image:
+    def __init__(self, scr, img, x, y):
+        self.scr = scr
+        self.img = img
+        self.og = img
+        self.x = x
+        self.y = y
+        # constant
+        self.angle = 90
+        return
+
+    def blit(self):
+        self.scr.blit(self.img, (self.x, self.y))
+        return
+
+    def get_rekt(self):
+        _, _, w, h = self.img.get_rect()
+        return self.x, self.y, w, h
+
+    def rotate(self, angle):
+        self.img, self.rect = self.rot_center(self.og, self.angle)
+        return
+
+    def rotate_up(self, factor):
+        self.angle += factor
+        self.rotate(self.angle)
+        return
+
+    def rotate_down(self, factor):
+        self.rotate_up(-factor)
+        return
+
+    def rot_center(self, image, angle):
+
+        center = image.get_rect().center
+        rotated_image = pygame.transform.rotate(image, angle)
+        new_rect = rotated_image.get_rect(center = center)
+
+        return rotated_image, new_rect
+    pass
+
+class Image_Button(Button):
+    def __init__(self, scr, img, x, y, resp, delay_time=0.3):
+        self.img = Image(scr, img, x, y)
+        self.rect = self.img.get_rekt()
+        self.delay_time = delay_time
+        self.resp = resp
+        self.rotate_up = self.img.rotate_up
+        self.rotate_down = self.img.rotate_down
+        return
+
+    def blit(self):
+        self.img.blit()
+        super().blit()
+        return
+    pass
+