+++ sprite_manager.py
@@ -9,7 +9,7 @@ def get_sprites():
 	files = [f"sprites\\{file}" for file in files]
 	ratio_dividers = {"coin": 12, "dm_ex": 4, 'dm_paper': 3,
 	'dollar': 12,'handle_active': 3,'handle_idle': 3,
-	'letter': 12,'machine_base': 3,'ms_ex': 4,'ms_paper': 3}
+	'letter': 12,'machine_base': 2,'ms_ex': 4,'ms_paper': 3}
 	# single best for loop ever
 	for name, file in zip(names, files):
 		img = pygame.image.load(file)